requirements.txt
astroid==2.2.5
Django==2.1.7
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-redis==4.10.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==6.0.0
pycodestyle==2.5.0
pylint==2.3.1
PyMySQL==0.9.3
redis==3.2.1
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
python-docx==0.8.10 